@@ -76,34 +75,35 @@
         "@types/react-datepicker": "4.3.4",
         "@types/react-dom": "17.0.16",
         "@types/tinycolor2": "1.4.3",
-        "@vitejs/plugin-react-refresh": "1.3.6",
-        "autoprefixer": "10.4.4",
+        "autoprefixer": "10.4.7",
         "chalk": "5.0.1",
         "concat": "1.0.3",
         "cross-env": "7.0.3",
-        "cypress": "9.5.3",
-        "cypress-real-events": "1.7.0",
-        "eslint": "8.12.0",
-        "eslint-plugin-jsx-a11y": "6.5.1",
-        "esno": "0.14.1",
+        "cypress": "^10.3.1",
+        "cypress-real-events": "^1.7.1",
+        "eslint": "8.20.0",
+        "eslint-plugin-jsx-a11y": "6.6.1",
+        "eslint-plugin-storybook": "^0.6.1",
+        "esno": "0.16.3",
         "fast-glob": "3.2.11",
-        "postcss": "8.4.12",
+        "postcss": "8.4.14",
         "postcss-cli": "9.1.0",
         "postcss-import": "14.1.0",
         "postcss-nested": "5.0.6",
-        "prettier": "2.6.2",
+        "prettier": "2.7.1",
         "react": "17.0.2",
         "react-dom": "17.0.2",
         "rimraf": "3.0.2",
         "storybook-dark-mode": "1.1.0",
-        "tailwindcss": "3.1.4",
-        "ts-node": "10.7.0",
-        "typescript": "4.6.3",
-        "vite": "2.9.1",
-        "vite-plugin-dts": "1.0.5"
+        "tailwindcss": "3.1.6",
+        "ts-node": "10.9.1",
+        "typescript": "4.7.4",
+        "vite": "^3.0.3",
+        "vite-plugin-dts": "1.4.0",
+        "vite-plugin-eslint": "^1.7.0"
     },
     "dependencies": {
-        "@frontify/fondue-tokens": "^3.0.1",
+        "@frontify/fondue-tokens": "^3.1.0",
         "@popperjs/core": "^2.11.5",
         "@react-aria/accordion": "3.0.0-alpha.9",
         "@react-aria/aria-modal-polyfill": "^3.4.3",
