@@ -1,5 +1,5 @@
 module.exports = {
-    extends: ['@frontify/eslint-config-react', 'plugin:jsx-a11y/recommended'],
+    extends: ['@frontify/eslint-config-react', 'plugin:jsx-a11y/recommended', 'plugin:storybook/recommended'],
     settings: {
         react: {
             version: 'detect',
