@@ -1,5 +1,5 @@
 import { LinkElement } from '@components/RichTextEditor/components';
-import { createPluginFactory, ELEMENT_LINK } from '@udecode/plate';
+import { ELEMENT_LINK, createPluginFactory } from '@udecode/plate';
 
 // Extends the default link plugin so that 'getChosenLink()' can be given into
 // https://github.com/udecode/plate/tree/main/packages/nodes/link/src
