@@ -130,7 +130,7 @@ export const mockNodesFlat: mockNodeType = () => [
         label: 'Document',
         value: 'https://weare.frontify.com/document/1',
         icon: <IconFolder size={IconSize.Size16} />,
-        actions: [<Button size={ButtonSize.Small} icon={<IconPlus></IconPlus>} />],
+        actions: [<Button key={'Design System Testing'} size={ButtonSize.Small} icon={<IconPlus></IconPlus>} />],
         parentId: null,
         sort: 1,
         editable: true,
@@ -139,7 +139,13 @@ export const mockNodesFlat: mockNodeType = () => [
         id: '1-1',
         name: 'Uncategorizes Pages',
         icon: <IconFolder size={IconSize.Size16} />,
-        actions: [<Button size={ButtonSize.Small} icon={<IconPlus></IconPlus>} />],
+        actions: [
+            <Button
+                key={'Design System Testing Uncategorizes Pages'}
+                size={ButtonSize.Small}
+                icon={<IconPlus></IconPlus>}
+            />,
+        ],
         parentId: '1',
         sort: 1,
     },
@@ -150,7 +156,13 @@ export const mockNodesFlat: mockNodeType = () => [
         name: 'Test Category',
         label: 'Document',
         icon: <IconFolder size={IconSize.Size16} />,
-        actions: [<Button size={ButtonSize.Small} icon={<IconPlus></IconPlus>} />],
+        actions: [
+            <Button
+                key={'Design System Testing Uncategorizes Pages Test'}
+                size={ButtonSize.Small}
+                icon={<IconPlus></IconPlus>}
+            />,
+        ],
         value: 'https://weare.frontify.com/document/923#/test',
         sort: null,
         badge: <IconSmileysPeople size={IconSize.Size16} />,
@@ -160,7 +172,13 @@ export const mockNodesFlat: mockNodeType = () => [
         id: '1-3',
         name: 'Category Without Subitems',
         icon: <IconFolder size={IconSize.Size16} />,
-        actions: [<Button size={ButtonSize.Small} icon={<IconPlus></IconPlus>} />],
+        actions: [
+            <Button
+                key={'Design System Testing Uncategorizes Pages Test Subitems'}
+                size={ButtonSize.Small}
+                icon={<IconPlus></IconPlus>}
+            />,
+        ],
         parentId: '1',
         value: 'https://weare.frontify.com/document/923#/test',
         sort: null,
