@@ -12,18 +12,21 @@ function IconArrowAlignHorizontalCentre12(props: GeneratedIconProps): React.Reac
             name="IconArrowAlignHorizontalCentre12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M6.5.5a.5.5 0 0 0-1 0v2.663L4.233 1.896a.5.5 0 1 0-.707.708l2.122 2.12.353.354.354-.353 2.121-2.121a.5.5 0 1 0-.707-.708L6.5 3.166V.5ZM1 6a.5.5 0 0 1 .5-.5h9a.5.5 0 0 1 0 1h-9A.5.5 0 0 1 1 6Zm5.002 6.078a.5.5 0 0 1-.5-.5V8.913l-1.269 1.269a.5.5 0 0 1-.707-.707l2.122-2.121L6 7l.354.354 2.121 2.12a.5.5 0 0 1-.707.708L6.502 8.915v2.663a.5.5 0 0 1-.5.5Z"
-                clipRule="evenodd"
-            />
-
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M6.5.5a.5.5 0 0 0-1 0v2.663L4.233 1.896a.5.5 0 1 0-.707.708l2.122 2.12.353.354.354-.353 2.121-2.121a.5.5 0 1 0-.707-.708L6.5 3.166V.5ZM1 6a.5.5 0 0 1 .5-.5h9a.5.5 0 0 1 0 1h-9A.5.5 0 0 1 1 6Zm5.002 6.078a.5.5 0 0 1-.5-.5V8.913l-1.269 1.269a.5.5 0 0 1-.707-.707l2.122-2.121L6 7l.354.354 2.121 2.12a.5.5 0 0 1-.707.708L6.502 8.915v2.663a.5.5 0 0 1-.5.5Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
