@@ -12,17 +12,21 @@ function IconBracketsCurly12(props: GeneratedIconProps): React.ReactElement<Gene
             name="IconBracketsCurly12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M6.996.473a.5.5 0 0 0 .464.526l.928 1.037A3.7 3.7 0 0 1 8.5 3c0 1.608.657 2.517 1.401 3.002C9.157 6.485 8.5 7.393 8.5 9c0 1.098-.331 1.572-.568 1.779a.979.979 0 0 1-.431.22.5.5 0 1 0 .053 1l-.03-.499.03.5.001-.001h.006l.01-.001a1.072 1.072 0 0 0 .103-.014c.061-.01.143-.03.238-.06.19-.064.436-.18.68-.393.501-.44.908-1.216.908-2.531 0-1.332.536-1.914 1.002-2.194a2.192 2.192 0 0 1 .977-.301h.011-.002a.5.5 0 0 0 .004-1L9.509 3.288l1.983 2.215h-.011l-.056-.004a2.201 2.201 0 0 1-.922-.3c-.432-.262-.923-.782-.994-1.911l-1.12-1.253 1.12 1.252A4.568 4.568 0 0 1 9.5 3c0-1.32-.416-2.098-.927-2.536a2.005 2.005 0 0 0-.689-.39A1.727 1.727 0 0 0 7.54.003h-.01L7.526 0h-.002l-.029.5.027-.5a.5.5 0 0 0-.527.473Zm4.495 5.031ZM8.388 2.036 7.46 1h.017a1.006 1.006 0 0 1 .445.221c.154.133.347.373.466.814ZM4.988.473A.5.5 0 0 1 4.507 1a1.006 1.006 0 0 0-.445.221c-.154.133-.347.373-.466.814L2.475 3.288A4.43 4.43 0 0 0 2.484 3c0-1.32.416-2.098.926-2.536.248-.213.498-.327.69-.39a1.727 1.727 0 0 1 .344-.072h.01L4.458 0h.002l.028.5-.027-.5a.5.5 0 0 1 .527.473ZM2.475 3.289.492 5.503M2.475 3.29.492 5.504m0 0h.01L.56 5.5a2.201 2.201 0 0 0 .922-.3c.431-.262.922-.782.994-1.911l1.12-1.253A3.7 3.7 0 0 0 3.485 3c0 1.608-.658 2.517-1.401 3.002.744.483 1.4 1.391 1.4 2.998 0 1.098.332 1.572.568 1.779a.979.979 0 0 0 .432.22.5.5 0 1 1-.053 1l.03-.499-.03.5-.001-.001h-.006l-.01-.001a1.072 1.072 0 0 1-.103-.014 1.687 1.687 0 0 1-.238-.06 1.978 1.978 0 0 1-.68-.393c-.501-.44-.908-1.216-.908-2.531 0-1.332-.536-1.914-1.002-2.194a2.193 2.193 0 0 0-.978-.301h-.01.001a.5.5 0 0 1-.003-1m0 0Zm0 0h.002-.001Zm3.104-3.469L4.524 1l-.928 1.037Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M6.996.473a.5.5 0 0 0 .464.526l.928 1.037A3.7 3.7 0 0 1 8.5 3c0 1.608.657 2.517 1.401 3.002C9.157 6.485 8.5 7.393 8.5 9c0 1.098-.331 1.572-.568 1.779a.979.979 0 0 1-.431.22.5.5 0 1 0 .053 1l-.03-.499.03.5.001-.001h.006l.01-.001a1.072 1.072 0 0 0 .103-.014c.061-.01.143-.03.238-.06.19-.064.436-.18.68-.393.501-.44.908-1.216.908-2.531 0-1.332.536-1.914 1.002-2.194a2.192 2.192 0 0 1 .977-.301h.011-.002a.5.5 0 0 0 .004-1L9.509 3.288l1.983 2.215h-.011l-.056-.004a2.201 2.201 0 0 1-.922-.3c-.432-.262-.923-.782-.994-1.911l-1.12-1.253 1.12 1.252A4.568 4.568 0 0 1 9.5 3c0-1.32-.416-2.098-.927-2.536a2.005 2.005 0 0 0-.689-.39A1.727 1.727 0 0 0 7.54.003h-.01L7.526 0h-.002l-.029.5.027-.5a.5.5 0 0 0-.527.473Zm4.495 5.031ZM8.388 2.036 7.46 1h.017a1.006 1.006 0 0 1 .445.221c.154.133.347.373.466.814ZM4.988.473A.5.5 0 0 1 4.507 1a1.006 1.006 0 0 0-.445.221c-.154.133-.347.373-.466.814L2.475 3.288A4.43 4.43 0 0 0 2.484 3c0-1.32.416-2.098.926-2.536.248-.213.498-.327.69-.39a1.727 1.727 0 0 1 .344-.072h.01L4.458 0h.002l.028.5-.027-.5a.5.5 0 0 1 .527.473ZM2.475 3.289.492 5.503M2.475 3.29.492 5.504m0 0h.01L.56 5.5a2.201 2.201 0 0 0 .922-.3c.431-.262.922-.782.994-1.911l1.12-1.253A3.7 3.7 0 0 0 3.485 3c0 1.608-.658 2.517-1.401 3.002.744.483 1.4 1.391 1.4 2.998 0 1.098.332 1.572.568 1.779a.979.979 0 0 0 .432.22.5.5 0 1 1-.053 1l.03-.499-.03.5-.001-.001h-.006l-.01-.001a1.072 1.072 0 0 1-.103-.014 1.687 1.687 0 0 1-.238-.06 1.978 1.978 0 0 1-.68-.393c-.501-.44-.908-1.216-.908-2.531 0-1.332-.536-1.914-1.002-2.194a2.193 2.193 0 0 0-.978-.301h-.01.001a.5.5 0 0 1-.003-1m0 0Zm0 0h.002-.001Zm3.104-3.469L4.524 1l-.928 1.037Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
