@@ -12,18 +12,21 @@ function IconCross16(props: GeneratedIconProps): React.ReactElement<GeneratedIco
             name="IconCross16"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M4.464 3.757a.5.5 0 0 0-.707.707L7.293 8l-3.536 3.536a.5.5 0 1 0 .707.707L8 8.707l3.536 3.536a.5.5 0 1 0 .707-.707L8.707 8l3.536-3.536a.5.5 0 0 0-.707-.707L8 7.293 4.464 3.757Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h16v16H0z" />
-                </clipPath>
-            </defs>
-
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M4.464 3.757a.5.5 0 0 0-.707.707L7.293 8l-3.536 3.536a.5.5 0 1 0 .707.707L8 8.707l3.536 3.536a.5.5 0 1 0 .707-.707L8.707 8l3.536-3.536a.5.5 0 0 0-.707-.707L8 7.293 4.464 3.757Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h16v16H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
