@@ -12,17 +12,21 @@ function IconCross24(props: GeneratedIconProps): React.ReactElement<GeneratedIco
             name="IconCross24"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="m11.647 13.06 5.303 5.304a.75.75 0 0 0 1.06-1.06L12.708 12l5.304-5.303a.75.75 0 0 0-1.06-1.061l-5.304 5.303-4.95-4.95a.75.75 0 0 0-1.06 1.061l4.95 4.95-4.95 4.95a.75.75 0 0 0 1.06 1.06l4.95-4.95Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h24v24H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="m11.647 13.06 5.303 5.304a.75.75 0 0 0 1.06-1.06L12.708 12l5.304-5.303a.75.75 0 0 0-1.06-1.061l-5.304 5.303-4.95-4.95a.75.75 0 0 0-1.06 1.061l4.95 4.95-4.95 4.95a.75.75 0 0 0 1.06 1.06l4.95-4.95Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h24v24H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
