@@ -12,17 +12,21 @@ function IconHeartCircleStack12(props: GeneratedIconProps): React.ReactElement<G
             name="IconHeartCircleStack12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M9.392 9.392a5 5 0 1 0-6.784-6.784 5 5 0 1 0 6.784 6.784Zm.56-1.692A4 4 0 0 0 4.3 2.049 5 5 0 0 1 9.951 7.7ZM5 11a4 4 0 1 0 0-8 4 4 0 0 0 0 8ZM3.377 6.023c-.246.246-.261.43-.213.635.068.287.29.655.662 1.065a8.26 8.26 0 0 0 1.129 1.006L5 8.763l.045-.034a8.26 8.26 0 0 0 1.13-1.006c.372-.41.593-.778.66-1.065.049-.206.034-.39-.212-.635-.176-.176-.278-.192-.3-.194-.031-.004-.085-.001-.175.043a1.29 1.29 0 0 0-.359.292 1 1 0 0 1-1.592-.016 1.29 1.29 0 0 0-.345-.276c-.09-.044-.144-.047-.176-.043-.02.002-.123.018-.3.194ZM5.79 4.937c-.477.202-.79.613-.79.613s-.313-.411-.79-.613c-.42-.178-.968-.194-1.54.379C.672 7.313 5 9.976 5 9.976s4.328-2.663 2.33-4.66c-.572-.573-1.12-.557-1.54-.38Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M9.392 9.392a5 5 0 1 0-6.784-6.784 5 5 0 1 0 6.784 6.784Zm.56-1.692A4 4 0 0 0 4.3 2.049 5 5 0 0 1 9.951 7.7ZM5 11a4 4 0 1 0 0-8 4 4 0 0 0 0 8ZM3.377 6.023c-.246.246-.261.43-.213.635.068.287.29.655.662 1.065a8.26 8.26 0 0 0 1.129 1.006L5 8.763l.045-.034a8.26 8.26 0 0 0 1.13-1.006c.372-.41.593-.778.66-1.065.049-.206.034-.39-.212-.635-.176-.176-.278-.192-.3-.194-.031-.004-.085-.001-.175.043a1.29 1.29 0 0 0-.359.292 1 1 0 0 1-1.592-.016 1.29 1.29 0 0 0-.345-.276c-.09-.044-.144-.047-.176-.043-.02.002-.123.018-.3.194ZM5.79 4.937c-.477.202-.79.613-.79.613s-.313-.411-.79-.613c-.42-.178-.968-.194-1.54.379C.672 7.313 5 9.976 5 9.976s4.328-2.663 2.33-4.66c-.572-.573-1.12-.557-1.54-.38Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
