@@ -12,17 +12,21 @@ function IconImageWithText12(props: GeneratedIconProps): React.ReactElement<Gene
             name="IconImageWithText12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M10.5 2h-9a.5.5 0 0 0-.5.5v7a.5.5 0 0 0 .5.5h1.293l2.854-2.854.205-.205.28.077L11 8.345V2.5a.5.5 0 0 0-.5-.5ZM12 8.982V2.5A1.5 1.5 0 0 0 10.5 1h-9A1.5 1.5 0 0 0 0 2.5v7A1.5 1.5 0 0 0 1.5 11h9A1.5 1.5 0 0 0 12 9.5v-.483a.51.51 0 0 0 0-.035Zm-1 .4L6.148 8.059 4.208 10H10.5a.5.5 0 0 0 .5-.5v-.118ZM8.5 5a.5.5 0 1 0 0-1 .5.5 0 0 0 0 1Zm0 1a1.5 1.5 0 1 0 0-3 1.5 1.5 0 0 0 0 3ZM2 3.5a.5.5 0 0 1 .5-.5h3a.5.5 0 0 1 0 1h-3a.5.5 0 0 1-.5-.5ZM2.5 5a.5.5 0 0 0 0 1h2a.5.5 0 0 0 0-1h-2Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M10.5 2h-9a.5.5 0 0 0-.5.5v7a.5.5 0 0 0 .5.5h1.293l2.854-2.854.205-.205.28.077L11 8.345V2.5a.5.5 0 0 0-.5-.5ZM12 8.982V2.5A1.5 1.5 0 0 0 10.5 1h-9A1.5 1.5 0 0 0 0 2.5v7A1.5 1.5 0 0 0 1.5 11h9A1.5 1.5 0 0 0 12 9.5v-.483a.51.51 0 0 0 0-.035Zm-1 .4L6.148 8.059 4.208 10H10.5a.5.5 0 0 0 .5-.5v-.118ZM8.5 5a.5.5 0 1 0 0-1 .5.5 0 0 0 0 1Zm0 1a1.5 1.5 0 1 0 0-3 1.5 1.5 0 0 0 0 3ZM2 3.5a.5.5 0 0 1 .5-.5h3a.5.5 0 0 1 0 1h-3a.5.5 0 0 1-.5-.5ZM2.5 5a.5.5 0 0 0 0 1h2a.5.5 0 0 0 0-1h-2Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
