@@ -12,17 +12,21 @@ function IconLeaf16(props: GeneratedIconProps): React.ReactElement<GeneratedIcon
             name="IconLeaf16"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="m6 9.3-1.212 1.213A4.417 4.417 0 0 1 5.12 4.64c.38-.38 1.027-.749 1.886-1.061.843-.307 1.811-.532 2.758-.67.947-.14 1.84-.187 2.524-.155a4.502 4.502 0 0 1 .853.111l.016.06c.043.183.079.448.095.793.033.685-.015 1.577-.154 2.524-.139.947-.364 1.915-.67 2.758-.313.859-.681 1.506-1.062 1.886a4.417 4.417 0 0 1-5.872.334l1.212-1.212h2.98a.5.5 0 1 0 0-1h-1.98l2.437-2.437a.5.5 0 1 0-.707-.707L7 8.3V6.64a.5.5 0 0 0-1 0V9.3ZM4.414 3.933a5.417 5.417 0 0 0-.337 7.29l-1.36 1.36a.5.5 0 1 0 .707.707l1.36-1.36a5.417 5.417 0 0 0 7.29-.337c2.116-2.115 2.654-8.612 1.803-9.463-.851-.85-7.347-.313-9.463 1.803Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h16v16H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="m6 9.3-1.212 1.213A4.417 4.417 0 0 1 5.12 4.64c.38-.38 1.027-.749 1.886-1.061.843-.307 1.811-.532 2.758-.67.947-.14 1.84-.187 2.524-.155a4.502 4.502 0 0 1 .853.111l.016.06c.043.183.079.448.095.793.033.685-.015 1.577-.154 2.524-.139.947-.364 1.915-.67 2.758-.313.859-.681 1.506-1.062 1.886a4.417 4.417 0 0 1-5.872.334l1.212-1.212h2.98a.5.5 0 1 0 0-1h-1.98l2.437-2.437a.5.5 0 1 0-.707-.707L7 8.3V6.64a.5.5 0 0 0-1 0V9.3ZM4.414 3.933a5.417 5.417 0 0 0-.337 7.29l-1.36 1.36a.5.5 0 1 0 .707.707l1.36-1.36a5.417 5.417 0 0 0 7.29-.337c2.116-2.115 2.654-8.612 1.803-9.463-.851-.85-7.347-.313-9.463 1.803Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h16v16H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
