@@ -12,17 +12,21 @@ function IconLeaf24(props: GeneratedIconProps): React.ReactElement<GeneratedIcon
             name="IconLeaf24"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="m8.986 13.939-1.258 1.258a5.398 5.398 0 0 1 .494-7.065c.45-.45 1.238-.905 2.318-1.298 1.055-.384 2.273-.667 3.466-.842 1.195-.176 2.317-.235 3.171-.194.393.018.696.057.909.102.045.212.083.516.102.909.04.854-.019 1.975-.194 3.17-.175 1.194-.458 2.411-.842 3.466-.393 1.08-.848 1.868-1.298 2.319a5.398 5.398 0 0 1-7.065.494l1.246-1.247h3.758a.75.75 0 1 0 0-1.5h-2.258l3-3a.75.75 0 1 0-1.06-1.061l-2.989 2.989v-1.83a.75.75 0 0 0-1.5 0v3.33ZM7.161 7.07a6.897 6.897 0 0 0-.502 9.195l-1.74 1.739a.75.75 0 0 0 1.062 1.06l1.739-1.739a6.898 6.898 0 0 0 9.194-.502c2.694-2.693 3.379-10.964 2.295-12.047-1.084-1.084-9.355-.399-12.048 2.294Zm11.155-1.106-.003-.001h.003Zm-.294-.292-.001-.003.001.003Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h24v24H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="m8.986 13.939-1.258 1.258a5.398 5.398 0 0 1 .494-7.065c.45-.45 1.238-.905 2.318-1.298 1.055-.384 2.273-.667 3.466-.842 1.195-.176 2.317-.235 3.171-.194.393.018.696.057.909.102.045.212.083.516.102.909.04.854-.019 1.975-.194 3.17-.175 1.194-.458 2.411-.842 3.466-.393 1.08-.848 1.868-1.298 2.319a5.398 5.398 0 0 1-7.065.494l1.246-1.247h3.758a.75.75 0 1 0 0-1.5h-2.258l3-3a.75.75 0 1 0-1.06-1.061l-2.989 2.989v-1.83a.75.75 0 0 0-1.5 0v3.33ZM7.161 7.07a6.897 6.897 0 0 0-.502 9.195l-1.74 1.739a.75.75 0 0 0 1.062 1.06l1.739-1.739a6.898 6.898 0 0 0 9.194-.502c2.694-2.693 3.379-10.964 2.295-12.047-1.084-1.084-9.355-.399-12.048 2.294Zm11.155-1.106-.003-.001h.003Zm-.294-.292-.001-.003.001.003Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h24v24H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
