@@ -12,17 +12,21 @@ function IconMediaObjectRatio2To112(props: GeneratedIconProps): React.ReactEleme
             name="IconMediaObjectRatio2To112"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M1.5 2h4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-.5.5h-4a.5.5 0 0 1-.5-.5v-7a.5.5 0 0 1 .5-.5ZM0 2.5A1.5 1.5 0 0 1 1.5 1h4A1.5 1.5 0 0 1 7 2.5v7A1.5 1.5 0 0 1 5.5 11h-4A1.5 1.5 0 0 1 0 9.5v-7Zm9 0a.5.5 0 0 1 .5-.5h2.002a.5.5 0 1 1 0 1H9.5a.5.5 0 0 1-.5-.5Zm0 4a.5.5 0 0 1 .5-.5h2.002a.5.5 0 1 1 0 1H9.5a.5.5 0 0 1-.5-.5ZM9.5 4a.5.5 0 0 0 0 1h1a.5.5 0 0 0 0-1h-1ZM9 8.5a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 0 1h-1a.5.5 0 0 1-.5-.5Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M1.5 2h4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-.5.5h-4a.5.5 0 0 1-.5-.5v-7a.5.5 0 0 1 .5-.5ZM0 2.5A1.5 1.5 0 0 1 1.5 1h4A1.5 1.5 0 0 1 7 2.5v7A1.5 1.5 0 0 1 5.5 11h-4A1.5 1.5 0 0 1 0 9.5v-7Zm9 0a.5.5 0 0 1 .5-.5h2.002a.5.5 0 1 1 0 1H9.5a.5.5 0 0 1-.5-.5Zm0 4a.5.5 0 0 1 .5-.5h2.002a.5.5 0 1 1 0 1H9.5a.5.5 0 0 1-.5-.5ZM9.5 4a.5.5 0 0 0 0 1h1a.5.5 0 0 0 0-1h-1ZM9 8.5a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 0 1h-1a.5.5 0 0 1-.5-.5Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
