@@ -12,17 +12,21 @@ function IconMoney12(props: GeneratedIconProps): React.ReactElement<GeneratedIco
             name="IconMoney12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M5.5 0a.5.5 0 0 1 .5.5v.632c.939.047 1.65.417 2.131.784a3.956 3.956 0 0 1 .728.726L8.9 2.7l.012.018.004.006.002.002v.001S8.92 2.728 8.5 3l.419-.273a.5.5 0 0 1-.836.548L4.875 2.208l3.207 1.067-.003-.005a2.163 2.163 0 0 0-.124-.156 2.955 2.955 0 0 0-.431-.403 2.79 2.79 0 0 0-1.721-.584 4.87 4.87 0 0 0-.928.08l-1.44-.478c.42-.299.953-.476 1.565-.554V.5a.5.5 0 0 1 .5-.5ZM5 10.896a4.746 4.746 0 0 1-1.992-.755 4.13 4.13 0 0 1-.678-.57 3.088 3.088 0 0 1-.22-.257l-.014-.02-.005-.006-.001-.003h-.001v-.001L2.5 9l-.412.283a.5.5 0 0 1 .822-.57l.004.006.024.03c.023.03.062.076.116.133.11.114.279.272.514.43.466.315 1.196.641 2.235.641.664 0 1.197-.16 1.556-.452.342-.28.585-.727.585-1.45 0-.384-.192-.666-.618-.935-.442-.28-1.045-.485-1.733-.72l-.032-.01c-.666-.227-1.414-.484-1.99-.872C2.97 5.11 2.5 4.52 2.5 3.646c0-.81.313-1.475.936-1.917l1.439.479c-.37.072-.653.189-.86.336-.328.233-.515.577-.515 1.102 0 .443.21.756.63 1.039.445.3 1.058.517 1.753.754l.093.032c.636.216 1.34.455 1.884.8.595.376 1.084.93 1.084 1.78 0 .973-.342 1.726-.953 2.225-.546.445-1.254.643-1.99.673v.648a.5.5 0 0 1-1.001 0v-.7Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M5.5 0a.5.5 0 0 1 .5.5v.632c.939.047 1.65.417 2.131.784a3.956 3.956 0 0 1 .728.726L8.9 2.7l.012.018.004.006.002.002v.001S8.92 2.728 8.5 3l.419-.273a.5.5 0 0 1-.836.548L4.875 2.208l3.207 1.067-.003-.005a2.163 2.163 0 0 0-.124-.156 2.955 2.955 0 0 0-.431-.403 2.79 2.79 0 0 0-1.721-.584 4.87 4.87 0 0 0-.928.08l-1.44-.478c.42-.299.953-.476 1.565-.554V.5a.5.5 0 0 1 .5-.5ZM5 10.896a4.746 4.746 0 0 1-1.992-.755 4.13 4.13 0 0 1-.678-.57 3.088 3.088 0 0 1-.22-.257l-.014-.02-.005-.006-.001-.003h-.001v-.001L2.5 9l-.412.283a.5.5 0 0 1 .822-.57l.004.006.024.03c.023.03.062.076.116.133.11.114.279.272.514.43.466.315 1.196.641 2.235.641.664 0 1.197-.16 1.556-.452.342-.28.585-.727.585-1.45 0-.384-.192-.666-.618-.935-.442-.28-1.045-.485-1.733-.72l-.032-.01c-.666-.227-1.414-.484-1.99-.872C2.97 5.11 2.5 4.52 2.5 3.646c0-.81.313-1.475.936-1.917l1.439.479c-.37.072-.653.189-.86.336-.328.233-.515.577-.515 1.102 0 .443.21.756.63 1.039.445.3 1.058.517 1.753.754l.093.032c.636.216 1.34.455 1.884.8.595.376 1.084.93 1.084 1.78 0 .973-.342 1.726-.953 2.225-.546.445-1.254.643-1.99.673v.648a.5.5 0 0 1-1.001 0v-.7Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
