@@ -12,17 +12,21 @@ function IconPaintbrush24(props: GeneratedIconProps): React.ReactElement<Generat
             name="IconPaintbrush24"
             {...props}
         >
-                <path
-                    fill="currentColor"
-                    fillRule="evenodd"
-                    d="m20.141 6.339-2.2 2.2-2.278-2.279 2.2-2.2a1.611 1.611 0 0 1 2.277 2.28Zm.04 4.439-1.179-1.179 2.2-2.2a3.111 3.111 0 1 0-4.4-4.4l-2.2 2.2-1.178-1.178a3 3 0 0 0-4.243 0L7.414 5.789a1.5 1.5 0 0 0-.379 1.485 1.52 1.52 0 0 0-.093.086l-4.478 4.478a1.5 1.5 0 0 0 0 2.122l7.778 7.778a1.5 1.5 0 0 0 2.121 0l4.479-4.479c.03-.03.058-.06.085-.093a1.5 1.5 0 0 0 1.486-.378l1.768-1.768a3 3 0 0 0 0-4.242Zm-6.679-4.479.04-.039-1.179-1.178a1.5 1.5 0 0 0-2.121 0L8.474 6.849l.263.263c.117.066.227.149.327.248l7.778 7.778c.099.1.181.21.247.326l.263.263 1.768-1.767a1.5 1.5 0 0 0 0-2.122l-1.179-1.178-.039.039-1.06-1.06-2.279-2.279-1.06-1.06ZM7.964 8.46l-4.44 4.439 1.16 1.159 1.964-1.964a.75.75 0 0 1 1.06 1.06l-1.964 1.964 1.14 1.14.864-.864a.75.75 0 0 1 1.06 1.06l-.864.864 1.14 1.14 2.513-2.515a.75.75 0 0 1 1.061 1.061l-2.514 2.514 1.159 1.159 4.439-4.439L7.964 8.46Z"
-                    clipRule="evenodd"
-                />
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="m20.141 6.339-2.2 2.2-2.278-2.279 2.2-2.2a1.611 1.611 0 0 1 2.277 2.28Zm.04 4.439-1.179-1.179 2.2-2.2a3.111 3.111 0 1 0-4.4-4.4l-2.2 2.2-1.178-1.178a3 3 0 0 0-4.243 0L7.414 5.789a1.5 1.5 0 0 0-.379 1.485 1.52 1.52 0 0 0-.093.086l-4.478 4.478a1.5 1.5 0 0 0 0 2.122l7.778 7.778a1.5 1.5 0 0 0 2.121 0l4.479-4.479c.03-.03.058-.06.085-.093a1.5 1.5 0 0 0 1.486-.378l1.768-1.768a3 3 0 0 0 0-4.242Zm-6.679-4.479.04-.039-1.179-1.178a1.5 1.5 0 0 0-2.121 0L8.474 6.849l.263.263c.117.066.227.149.327.248l7.778 7.778c.099.1.181.21.247.326l.263.263 1.768-1.767a1.5 1.5 0 0 0 0-2.122l-1.179-1.178-.039.039-1.06-1.06-2.279-2.279-1.06-1.06ZM7.964 8.46l-4.44 4.439 1.16 1.159 1.964-1.964a.75.75 0 0 1 1.06 1.06l-1.964 1.964 1.14 1.14.864-.864a.75.75 0 0 1 1.06 1.06l-.864.864 1.14 1.14 2.513-2.515a.75.75 0 0 1 1.061 1.061l-2.514 2.514 1.159 1.159 4.439-4.439L7.964 8.46Z"
+                        clipRule="evenodd"
+                    />
+                </g>
                 <defs>
                     <clipPath id="prefix__a">
                         <path fill="#fff" d="M0 0h24v24H0z" />
                     </clipPath>
                 </defs>
+            </g>
         </svg>
     );
 }
