@@ -12,17 +12,21 @@ function IconSpeechBubbleStack12(props: GeneratedIconProps): React.ReactElement<
             name="IconSpeechBubbleStack12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M3 7.5V6.395l-1.1.11-.35.035A.5.5 0 0 1 1 6.043V2.5a.5.5 0 0 1 .5-.5h6a.5.5 0 0 1 .5.5v3.7a.5.5 0 0 1-.5.5H4.498l-.296.314L3 8.285V7.5Zm4.5.2H4.929L3.727 8.973c-.57.603-1.545.284-1.705-.47A1.047 1.047 0 0 1 2 8.286V7.5l-.35.035A1.498 1.498 0 0 1 0 6.043V2.5A1.5 1.5 0 0 1 1.5 1h6A1.5 1.5 0 0 1 9 2.5v3.7a1.5 1.5 0 0 1-1.5 1.5ZM6.858 10H4.284l1.04-1h1.912l.283.25 1.257 1.107V9h1.745a.5.5 0 0 0 .5-.5v-3a.5.5 0 0 0-.5-.5H10V4h.521a1.5 1.5 0 0 1 1.5 1.5v3a1.5 1.5 0 0 1-1.5 1.5h-.745v.911a.75.75 0 0 1-1.246.563L6.858 10Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M3 7.5V6.395l-1.1.11-.35.035A.5.5 0 0 1 1 6.043V2.5a.5.5 0 0 1 .5-.5h6a.5.5 0 0 1 .5.5v3.7a.5.5 0 0 1-.5.5H4.498l-.296.314L3 8.285V7.5Zm4.5.2H4.929L3.727 8.973c-.57.603-1.545.284-1.705-.47A1.047 1.047 0 0 1 2 8.286V7.5l-.35.035A1.498 1.498 0 0 1 0 6.043V2.5A1.5 1.5 0 0 1 1.5 1h6A1.5 1.5 0 0 1 9 2.5v3.7a1.5 1.5 0 0 1-1.5 1.5ZM6.858 10H4.284l1.04-1h1.912l.283.25 1.257 1.107V9h1.745a.5.5 0 0 0 .5-.5v-3a.5.5 0 0 0-.5-.5H10V4h.521a1.5 1.5 0 0 1 1.5 1.5v3a1.5 1.5 0 0 1-1.5 1.5h-.745v.911a.75.75 0 0 1-1.246.563L6.858 10Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
