@@ -12,17 +12,21 @@ function IconStrikethroughBox12(props: GeneratedIconProps): React.ReactElement<G
             name="IconStrikethroughBox12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M11.854.854a.5.5 0 0 0-.708-.708l-1.13 1.13A1.99 1.99 0 0 0 9 1H3a2 2 0 0 0-2 2v6c0 .37.101.718.277 1.016l-1.13 1.13a.5.5 0 0 0 .707.708l1.13-1.13A2 2 0 0 0 3 11h6a2 2 0 0 0 2-2V3c0-.37-.101-.718-.277-1.016l1.13-1.13Zm-2.595 1.18A1.001 1.001 0 0 0 9 2H3a1 1 0 0 0-1 1v6c0 .09.012.176.034.259l7.225-7.225ZM2.741 9.966l7.225-7.225c.022.083.034.17.034.259v6a1 1 0 0 1-1 1H3c-.09 0-.176-.012-.259-.034Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M11.854.854a.5.5 0 0 0-.708-.708l-1.13 1.13A1.99 1.99 0 0 0 9 1H3a2 2 0 0 0-2 2v6c0 .37.101.718.277 1.016l-1.13 1.13a.5.5 0 0 0 .707.708l1.13-1.13A2 2 0 0 0 3 11h6a2 2 0 0 0 2-2V3c0-.37-.101-.718-.277-1.016l1.13-1.13Zm-2.595 1.18A1.001 1.001 0 0 0 9 2H3a1 1 0 0 0-1 1v6c0 .09.012.176.034.259l7.225-7.225ZM2.741 9.966l7.225-7.225c.022.083.034.17.034.259v6a1 1 0 0 1-1 1H3c-.09 0-.176-.012-.259-.034Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
