@@ -12,17 +12,21 @@ function IconTextBoxStack12(props: GeneratedIconProps): React.ReactElement<Gener
             name="IconTextBoxStack12"
             {...props}
         >
-            <path
-                fill="currentColor"
-                fillRule="evenodd"
-                d="M10.5 3h-7a.5.5 0 0 0-.5.5V4h5.5A1.5 1.5 0 0 1 10 5.5V8h.5a.5.5 0 0 0 .5-.5v-4a.5.5 0 0 0-.5-.5ZM10 9h.5A1.5 1.5 0 0 0 12 7.5v-4A1.5 1.5 0 0 0 10.5 2h-7A1.5 1.5 0 0 0 2 3.5V4h-.5A1.5 1.5 0 0 0 0 5.5v4A1.5 1.5 0 0 0 1.5 11h7A1.5 1.5 0 0 0 10 9.5V9ZM1.5 5h7a.5.5 0 0 1 .5.5v4a.5.5 0 0 1-.5.5h-7a.5.5 0 0 1-.5-.5v-4a.5.5 0 0 1 .5-.5Zm1 1.5A.5.5 0 0 1 3 6h4.286a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5ZM3 8a.5.5 0 0 0 0 1h2.571a.5.5 0 1 0 0-1H3Z"
-                clipRule="evenodd"
-            />
-            <defs>
-                <clipPath id="prefix__a">
-                    <path fill="#fff" d="M0 0h12v12H0z" />
-                </clipPath>
-            </defs>
+            <g fill="none">
+                <g clipPath="url(#prefix__a)">
+                    <path
+                        fill="currentColor"
+                        fillRule="evenodd"
+                        d="M10.5 3h-7a.5.5 0 0 0-.5.5V4h5.5A1.5 1.5 0 0 1 10 5.5V8h.5a.5.5 0 0 0 .5-.5v-4a.5.5 0 0 0-.5-.5ZM10 9h.5A1.5 1.5 0 0 0 12 7.5v-4A1.5 1.5 0 0 0 10.5 2h-7A1.5 1.5 0 0 0 2 3.5V4h-.5A1.5 1.5 0 0 0 0 5.5v4A1.5 1.5 0 0 0 1.5 11h7A1.5 1.5 0 0 0 10 9.5V9ZM1.5 5h7a.5.5 0 0 1 .5.5v4a.5.5 0 0 1-.5.5h-7a.5.5 0 0 1-.5-.5v-4a.5.5 0 0 1 .5-.5Zm1 1.5A.5.5 0 0 1 3 6h4.286a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5ZM3 8a.5.5 0 0 0 0 1h2.571a.5.5 0 1 0 0-1H3Z"
+                        clipRule="evenodd"
+                    />
+                </g>
+                <defs>
+                    <clipPath id="prefix__a">
+                        <path fill="#fff" d="M0 0h12v12H0z" />
+                    </clipPath>
+                </defs>
+            </g>
         </svg>
     );
 }
