@@ -34,11 +34,16 @@ export default defineConfig({
     resolve: {
         alias,
     },
+    // needs to be defined here, such that it is not undefined in the tests.
+    define: {
+        'process.env.REACT_APP_SC_ATTR': JSON.stringify(process.env.REACT_APP_SC_ATTR),
+        'process.env.SC_ATTR': JSON.stringify(process.env.SC_ATTR),
+    },
     plugins: [dts({ insertTypesEntry: true })],
     build: {
         lib: {
             entry: resolve(__dirname, 'src/index.ts'),
-            fileName: (format: string) => `index.${format}.js`,
+            fileName: (format: string) => `[name].${format}.js`,
             name: 'Fondue',
         },
         sourcemap: true,
